+++ sdk/identity/azure-identity/test/ut/azure_cli_credential_test.cpp
@@ -344,9 +344,68 @@ TEST(AzureCliCredential, UnsafeChars)
   }
 }
 
-TEST(AzureCliCredential, UnsafeCharsScopeAndTenantId)
+class ParameterizedTestForDisallowedChars : public ::testing::TestWithParam<std::string> {
+protected:
+  std::string value;
+};
+
+TEST_P(ParameterizedTestForDisallowedChars, DisallowedCharsForScopeAndTenantId)
 {
-  std::string const ValidScopeButNotTenantId = "string:withcolon";
+  std::string const InvalidValue = GetParam();
+
+  {
+    AzureCliCredentialOptions options;
+    options.TenantId = "01234567-89AB-CDEF-0123-456789ABCDEF";
+    options.TenantId += InvalidValue;
+    AzureCliCredential azCliCred(options);
+
+    TokenRequestContext trc;
+    trc.Scopes.push_back(std::string("https://storage.azure.com/.default"));
+    EXPECT_THROW(static_cast<void>(azCliCred.GetToken(trc, {})), AuthenticationException);
+
+    try
+    {
+      auto const token = azCliCred.GetToken(trc, {});
+    }
+    catch (AuthenticationException const& e)
+    {
+      EXPECT_TRUE(std::string(e.what()).find("Unsafe") != std::string::npos) << e.what();
+    }
+  }
+
+  {
+    AzureCliCredentialOptions options;
+    options.CliProcessTimeout = std::chrono::hours(24);
+    AzureCliCredential azCliCred(options);
+
+    TokenRequestContext trc;
+    trc.Scopes.push_back(std::string("https://storage.azure.com/.default") + InvalidValue);
+    EXPECT_THROW(static_cast<void>(azCliCred.GetToken(trc, {})), AuthenticationException);
+
+    try
+    {
+      auto const token = azCliCred.GetToken(trc, {});
+    }
+    catch (AuthenticationException const& e)
+    {
+      EXPECT_TRUE(std::string(e.what()).find("Unsafe") != std::string::npos) << e.what();
+    }
+  }
+}
+
+INSTANTIATE_TEST_SUITE_P(
+    AzureCliCredential,
+    ParameterizedTestForDisallowedChars,
+    ::testing::Values(" ", "|", "`", "\"", "'", ";", "&"));
+
+class ParameterizedTestForCharDifferences : public ::testing::TestWithParam<std::string> {
+protected:
+  std::string value;
+};
+
+TEST_P(ParameterizedTestForCharDifferences, ValidCharsForScopeButNotTenantId)
+{
+  std::string const ValidScopeButNotTenantId = GetParam();
 
   {
     AzureCliCredentialOptions options;
@@ -377,6 +436,7 @@ TEST(AzureCliCredential, UnsafeCharsScopeAndTenantId)
     trc.Scopes.push_back(
         std::string("https://storage.azure.com/.default") + ValidScopeButNotTenantId);
 
+    // We expect the GetToken to fail, but not because of the unsafe chars.
     try
     {
       auto const token = azCliCred.GetToken(trc, {});
@@ -387,6 +447,65 @@ TEST(AzureCliCredential, UnsafeCharsScopeAndTenantId)
     }
   }
 }
+
+INSTANTIATE_TEST_SUITE_P(
+    AzureCliCredential,
+    ParameterizedTestForCharDifferences,
+    ::testing::Values(":", "/", "_"));
+
+class ParameterizedTestForAllowedChars : public ::testing::TestWithParam<std::string> {
+protected:
+  std::string value;
+};
+
+TEST_P(ParameterizedTestForAllowedChars, ValidCharsForScopeAndTenantId)
+{
+  std::string const ValidChars = GetParam();
+
+  {
+    AzureCliCredentialOptions options;
+    options.TenantId = "01234567-89AB-CDEF-0123-456789ABCDEF";
+    options.TenantId += ValidChars;
+    AzureCliCredential azCliCred(options);
+
+    TokenRequestContext trc;
+    trc.Scopes.push_back(std::string("https://storage.azure.com/.default"));
+
+    // We expect the GetToken to fail, but not because of the unsafe chars.
+    try
+    {
+      auto const token = azCliCred.GetToken(trc, {});
+    }
+    catch (AuthenticationException const& e)
+    {
+      EXPECT_TRUE(std::string(e.what()).find("Unsafe") == std::string::npos) << e.what();
+    }
+  }
+
+  {
+    AzureCliCredentialOptions options;
+    options.CliProcessTimeout = std::chrono::hours(24);
+    AzureCliCredential azCliCred(options);
+
+    TokenRequestContext trc;
+    trc.Scopes.push_back(std::string("https://storage.azure.com/.default") + ValidChars);
+
+    // We expect the GetToken to fail, but not because of the unsafe chars.
+    try
+    {
+      auto const token = azCliCred.GetToken(trc, {});
+    }
+    catch (AuthenticationException const& e)
+    {
+      EXPECT_TRUE(std::string(e.what()).find("Unsafe") == std::string::npos) << e.what();
+    }
+  }
+}
+
+INSTANTIATE_TEST_SUITE_P(
+    AzureCliCredential,
+    ParameterizedTestForAllowedChars,
+    ::testing::Values(".", "-", "A", "9"));
 
 TEST(AzureCliCredential, StrictIso8601TimeFormat)
 {