+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Other Changes
 
-- Create separate lists of characters that are allowed within a tenant id vs scopes in `AzureCliCredential`.
+- Create separate lists of characters that are allowed within tenant ids and scopes in `AzureCliCredential`.
 
 ## 1.6.0-beta.3 (2023-10-12)